requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
requests==2.31.0
pydantic-settings==2.0.3
pytest==7.4.0
langchain==0.2.0
langgraph==0.0.24